// web-ui/client/gen/sales/lead.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */
import type { LeadStatusEnum } from './leadStatusEnum';

export interface Lead {
  readonly id: number;
  /** @maxLength 255 */
  name: string;
  /** @maxLength 254 */
  email: string;
  /** @maxLength 20 */
  phone?: string;
  /** @maxLength 255 */
  company?: string;
  /** @maxLength 100 */
  source?: string;
  status?: LeadStatusEnum;
  /** @nullable */
  assigned_to?: number | null;
  readonly assigned_to_name: string;
  readonly created_at: string;
  readonly updated_at: string;
}